fix: change kanban_cards.sort_order from integer to bigint
Date.now() produces values ~1.78 trillion in 2026, which overflows
PostgreSQL INTEGER (max 2,147,483,647). Changed to BIGINT in migration.
Also applied ALTER TABLE directly on production.

+++ backend/migrations/20260522000000_full_domain_schema.js
@@ -213,7 +213,7 @@ export const up = (pgm) => {
     id: { type: 'uuid', primaryKey: true, default: pgm.func('uuid_generate_v4()') },
     column_id: { type: 'uuid', notNull: true, references: 'kanban_columns', onDelete: 'CASCADE' },
     name: { type: 'varchar(255)', default: "''" },
-    sort_order: { type: 'integer', notNull: true, default: 0 },
+    sort_order: { type: 'bigint', notNull: true, default: 0 },
     template_id: { type: 'uuid' },
     fields: { type: 'jsonb', default: pgm.func("'{}'::jsonb") },
     contact_name: { type: 'varchar(255)', default: "''" },
